move error decoding to DecodeError and expose
