+++ subxt/src/tx/tx_progress.rs
@@ -11,8 +11,6 @@ use crate::{
     error::{
         DispatchError,
         Error,
-        ModuleError,
-        ModuleErrorData,
         TransactionError,
     },
     events::{
@@ -23,11 +21,9 @@ use crate::{
         Phase,
         StaticEvent,
     },
-    metadata::Metadata,
     rpc::SubstrateTxStatus,
     Config,
 };
-use codec::Decode;
 use derivative::Derivative;
 use futures::{
     Stream,
@@ -37,7 +33,6 @@ use jsonrpsee::core::{
     client::Subscription as RpcSubscription,
     Error as RpcError,
 };
-use scale_info::TypeDef;
 use sp_runtime::traits::Hash;
 
 pub use sp_runtime::traits::SignedExtension;
@@ -349,7 +344,7 @@ impl<T: Config, C: OnlineClientT<T>> TxInBlock<T, C> {
             let ev = ev?;
             if ev.pallet_name() == "System" && ev.variant_name() == "ExtrinsicFailed" {
                 let dispatch_error =
-                    decode_dispatch_error(&self.client.metadata(), ev.field_bytes());
+                    DispatchError::decode_from(ev.field_bytes(), &self.client.metadata());
                 return Err(dispatch_error.into())
             }
         }
@@ -458,109 +453,4 @@ impl<T: Config> TxEvents<T> {
     pub fn has<Ev: StaticEvent>(&self) -> Result<bool, Error> {
         Ok(self.find::<Ev>().next().transpose()?.is_some())
     }
-}
-
-// Attempt to decode a DispatchError, returning either the ModuleError it decodes to,
-// along with additional details on the error, or returning the raw bytes to work with
-// downstream if we couldn't decode it.
-fn decode_dispatch_error(metadata: &Metadata, bytes: &[u8]) -> DispatchError {
-    let dispatch_error_ty_id = match metadata.dispatch_error_ty() {
-        Some(id) => id,
-        None => {
-            tracing::warn!(
-                "Can't decode error: sp_runtime::DispatchError was not found in Metadata"
-            );
-            return DispatchError::Other(bytes.to_vec())
-        }
-    };
-
-    let dispatch_error_ty = match metadata.types().resolve(dispatch_error_ty_id) {
-        Some(ty) => ty,
-        None => {
-            tracing::warn!("Can't decode error: sp_runtime::DispatchError type ID doesn't resolve to a known type");
-            return DispatchError::Other(bytes.to_vec())
-        }
-    };
-
-    let variant = match dispatch_error_ty.type_def() {
-        TypeDef::Variant(var) => var,
-        _ => {
-            tracing::warn!(
-                "Can't decode error: sp_runtime::DispatchError type is not a Variant"
-            );
-            return DispatchError::Other(bytes.to_vec())
-        }
-    };
-
-    let module_variant_idx = variant
-        .variants()
-        .iter()
-        .find(|v| v.name() == "Module")
-        .map(|v| v.index());
-    let module_variant_idx = match module_variant_idx {
-        Some(idx) => idx,
-        None => {
-            tracing::warn!("Can't decode error: sp_runtime::DispatchError does not have a 'Module' variant");
-            return DispatchError::Other(bytes.to_vec())
-        }
-    };
-
-    // If the error bytes don't correspond to a ModuleError, just return the bytes.
-    // This is perfectly reasonable and expected, so no logging.
-    if bytes[0] != module_variant_idx {
-        return DispatchError::Other(bytes.to_vec())
-    }
-
-    // The remaining bytes are the module error, all being well:
-    let bytes = &bytes[1..];
-
-    // The oldest and second oldest type of error decode to this shape:
-    #[derive(Decode)]
-    struct LegacyModuleError {
-        index: u8,
-        error: u8,
-    }
-
-    // The newer case expands the error for forward compat:
-    #[derive(Decode)]
-    struct CurrentModuleError {
-        index: u8,
-        error: [u8; 4],
-    }
-
-    // try to decode into the new shape, or the old if that doesn't work
-    let err = match CurrentModuleError::decode(&mut &*bytes) {
-        Ok(e) => e,
-        Err(_) => {
-            let old_e = match LegacyModuleError::decode(&mut &*bytes) {
-                Ok(err) => err,
-                Err(_) => {
-                    tracing::warn!("Can't decode error: sp_runtime::DispatchError does not match known formats");
-                    return DispatchError::Other(bytes.to_vec())
-                }
-            };
-            CurrentModuleError {
-                index: old_e.index,
-                error: [old_e.error, 0, 0, 0],
-            }
-        }
-    };
-
-    let error_details = match metadata.error(err.index, err.error[0]) {
-        Ok(details) => details,
-        Err(_) => {
-            tracing::warn!("Can't decode error: sp_runtime::DispatchError::Module details do not match known information");
-            return DispatchError::Other(bytes.to_vec())
-        }
-    };
-
-    DispatchError::Module(ModuleError {
-        pallet: error_details.pallet().to_string(),
-        error: error_details.error().to_string(),
-        description: error_details.docs().to_vec(),
-        error_data: ModuleErrorData {
-            pallet_index: err.index,
-            error: err.error,
-        },
-    })
 }
